refactor: add extra aliases
