+++ tsconfig.json
@@ -17,7 +17,9 @@
     "jsx": "react-jsx",
     "baseUrl": "src",
     "paths": {
-      "components/*": ["components/*"]
+      "components/*": ["components/*"],
+      "data/*": ["data/*"],
+      "assets/*": ["assets/*"],
     }
   },
   "include": [
